requirements.txt
requests
python-dotenv
pandas
feedparser
firebase-admin
